# docs/agent-platform/agent-platform-openapi-v1.yaml
openapi: 3.1.0
info:
  title: Agent Platform Core API
  version: 1.4.0
  description: Core development contract for Agent, Release, Session, Run, Event, Approval and Artifact.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/agents/{agent_id}/publish:
    parameters:
      - $ref: '#/components/parameters/AgentId'
      - $ref: '#/components/parameters/IdempotencyKey'
    post:
      tags: [Releases]
      operationId: publishAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishAgentRequest'
      responses:
        '202':
          description: Release workflow accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseAccepted'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/agents/{agent_id}/publish-preview:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    post:
      tags: [Releases]
      operationId: previewAgentPublish
      description: Read-only publication preview; it never creates a version, Snapshot or Release.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishAgentPreviewRequest'
      responses:
        '200':
          description: Resolved Draft publication preview grouped by Runtime Target.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishAgentPreview'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/releases/{release_id}:
    parameters:
      - $ref: '#/components/parameters/ReleaseId'
    get:
      tags: [Releases]
      operationId: getRelease
      responses:
        '200':
          description: Release detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/agents/{agent_id}/rollback:
    parameters:
      - $ref: '#/components/parameters/AgentId'
      - $ref: '#/components/parameters/IdempotencyKey'
    post:
      tags: [Releases]
      operationId: rollbackAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollbackAgentRequest'
      responses:
        '202':
          description: Rollback release accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseAccepted'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/agents/{agent_id}/versions:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    get:
      tags: [Releases]
      operationId: listAgentVersions
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Agent version page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentVersionPage'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/agents/{agent_id}/versions/{version_id}:
    parameters:
      - $ref: '#/components/parameters/AgentId'
      - $ref: '#/components/parameters/VersionId'
    get:
      tags: [Releases]
      operationId: getAgentVersion
      responses:
        '200':
          description: Immutable Agent version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentVersion'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/agents/{agent_id}/diff:
    parameters:
      - $ref: '#/components/parameters/AgentId'
      - $ref: '#/components/parameters/FromSnapshotId'
      - $ref: '#/components/parameters/ToSnapshotId'
    get:
      tags: [Releases]
      operationId: diffAgentSnapshots
      responses:
        '200':
          description: Structured Snapshot diff used by publish and rollback review.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotDiff'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/deployments/{deployment_id}:
    parameters:
      - $ref: '#/components/parameters/DeploymentId'
    get:
      tags: [Releases]
      operationId: getDeployment
      responses:
        '200':
          description: Deployment detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceAuth:
      type: mutualTLS
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    AgentId:
      name: agent_id
      in: path
      required: true
      schema:
        type: string
    VersionId:
      name: version_id
      in: path
      required: true
      schema:
        type: string
    FromSnapshotId:
      name: from_snapshot_id
      in: query
      required: true
      schema:
        type: string
    ToSnapshotId:
      name: to_snapshot_id
      in: query
      required: true
      schema:
        type: string
...
    ReleaseId:
      name: release_id
      in: path
      required: true
      schema:
        type: string
    DeploymentId:
      name: deployment_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ErrorResponse:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
...
    PublishAgentRequest:
      type: object
      additionalProperties: false
      required: [expected_agent_version, runtime_targets, release_note]
      properties:
        expected_agent_version:
          type: integer
          minimum: 1
        runtime_targets:
          type: array
          minItems: 1
          items:
            type: string
        release_note:
          type: string
          minLength: 1
          maxLength: 2000
        run_smoke_test:
          type: boolean
          default: true
        activate_on_success:
          type: boolean
          default: true
    PublishAgentPreviewRequest:
      type: object
      additionalProperties: false
      required: [expected_agent_version, runtime_targets]
      properties:
        expected_agent_version:
          type: integer
          minimum: 1
        runtime_targets:
          type: array
          minItems: 1
          maxItems: 32
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 255
    PublishAgentPreview:
      type: object
      additionalProperties: false
      required: [agent_id, expected_agent_version, preview_snapshot_hash, resolved_bindings, targets, ready_to_publish]
      properties:
        agent_id:
          type: string
        expected_agent_version:
          type: integer
          minimum: 1
        preview_snapshot_hash:
          type: string
          pattern: '^sha256:[a-f0-9]{64}$'
        resolved_bindings:
          type: array
          items:
            $ref: '#/components/schemas/ResolvedPublishBinding'
        targets:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PublishAgentPreviewTarget'
        ready_to_publish:
          type: boolean
          description: True only when Snapshot prerequisites are present; Release supply-chain and deployment gates are evaluated again after submission.
    ResolvedPublishBinding:
      type: object
      additionalProperties: false
      required: [resource_type, resource_id, version_id, version_no, content_hash]
      properties:
        resource_type:
          type: string
          enum: [prompt, skill, mcp, model, sandbox, agent]
        resource_id:
          type: string
        version_id:
          type: string
        version_no:
          type: integer
          minimum: 1
        content_hash:
          type: string
          pattern: '^sha256:[a-f0-9]{64}$'
        binding_role:
          type: [string, 'null']
          enum: [primary, fallback_1, fallback_2, null]
    PublishAgentPreviewTarget:
      type: object
      additionalProperties: false
      required: [runtime_target_id, current_deployment_id, current_snapshot_id, changes]
      properties:
        runtime_target_id:
          type: string
        current_deployment_id:
          type: [string, 'null']
        current_snapshot_id:
          type: [string, 'null']
        changes:
          type: array
          items:
            $ref: '#/components/schemas/SnapshotDiffChangesItem'
    Release:
      type: object
      additionalProperties: false
      required: [id, agent_id, status, workflow_id, created_at]
      properties:
        id:
          type: string
        agent_id:
          type: string
        status:
          type: string
          enum: [REQUESTED, VALIDATING, COMPILING, SCANNING, SMOKE_TESTING, ACTIVATING, SUCCEEDED, FAILED, CANCELLED]
        workflow_id:
          type: string
        snapshot_id:
          type: [string, 'null']
        deployment_ids:
          type: array
          items:
            type: string
        error:
          anyOf:
            - $ref: '#/components/schemas/Error'
            - type: 'null'
        created_at:
          type: string
          format: date-time
    ReleaseAccepted:
      type: object
      additionalProperties: false
      required: [release_id, workflow_id, status, status_url]
      properties:
        release_id:
          type: string
        workflow_id:
          type: string
        status:
          type: string
          enum: [REQUESTED, VALIDATING]
        status_url:
          type: string
    RollbackAgentRequest:
      type: object
      additionalProperties: false
      required: [snapshot_id, runtime_targets, release_note]
      properties:
        snapshot_id:
          type: string
        runtime_targets:
          type: array
          minItems: 1
          items:
            type: string
        release_note:
          type: string
          minLength: 1
          maxLength: 2000
    AgentVersion:
      type: object
      additionalProperties: false
      required: [id, agent_id, version_no, snapshot_id, content_hash, created_at]
      properties:
        id:
          type: string
        agent_id:
          type: string
        version_no:
          type: integer
          minimum: 1
        snapshot_id:
          type: string
        content_hash:
          type: string
          pattern: '^sha256:[a-f0-9]{64}$'
        release_note:
          type: [string, 'null']
        created_at:
          type: string
          format: date-time
    AgentVersionPage:
      type: object
      additionalProperties: false
      required: [items, has_more]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AgentVersion'
        next_cursor:
          type: [string, 'null']
        has_more:
          type: boolean
    SnapshotDiff:
      type: object
      additionalProperties: false
      required: [agent_id, from_snapshot_id, to_snapshot_id, changes]
      properties:
        agent_id: { type: string }
        from_snapshot_id: { type: string }
        to_snapshot_id: { type: string }
        changes:
          type: array
          items:
            $ref: '#/components/schemas/SnapshotDiffChangesItem'
    SnapshotDiffChangesItem:
      type: object
      additionalProperties: false
      required: [category, path, change_type]
      properties:
        category: { enum: [resource_version, permission, network, sandbox, model, secret_reference, runtime] }
        path: { type: string }
        change_type: { enum: [added, removed, changed] }
        before: {}
        after: {}
        sensitive: { type: boolean, default: false }
...
    Deployment:
      type: object
      additionalProperties: false
      required: [id, agent_id, snapshot_id, bundle_id, runtime_target_id, status, compatibility_hash, created_at]
      properties:
        id:
          type: string
        agent_id:
          type: string
        snapshot_id:
          type: string
        bundle_id:
          type: string
        runtime_target_id:
          type: string
        status:
          type: string
          enum: [STAGED, ACTIVE, DEGRADED, RETIRED, FAILED]
        compatibility_hash:
          type: string
          pattern: '^sha256:[a-f0-9]{64}$'
        created_at:
          type: string
          format: date-time
        activated_at:
          type: [string, 'null']
          format: date-time
...